+++ lib/find.ts
@@ -1,13 +1,9 @@
-import curry from "./curry.ts";
-
 /**
  * **find** looks up first element of an array that matches the function passed
  * @param fn function that will be used to find an element, it has to return boolean
  * @param xs an array of elements to be searched
  * @returns first element that returned true
  */
-function find<T>(fn: (x: T) => boolean, xs: T[]): T {
+export default function find<T>(fn: (x: T) => boolean, xs: T[]): T {
     return xs.find(fn);
 }
-
-export default curry(find);