+++ lib/omit.ts
@@ -1,12 +1,10 @@
-import curry from "./curry.ts";
-
 /**
  * **omit** returns a copy of an object but without specified keys
  * @param keys keys to be omitted
  * @param object object to be copied
  * @returns an object without specified keys
  */
-function omit(keys: string[], object: object): object {
+export default function omit(keys: string[], object: object): object {
     let target = {};
 
     for (let i in object) {
@@ -18,5 +16,3 @@ function omit(keys: string[], object: object): object {
 
     return target;
 }
-
-export default curry(omit);